requirements/production.txt
-r base.txt
django-secure
django-csp
whitenoise[brotli] 